+++ crates/api/api/src/local_user/ban_person.rs
@@ -9,7 +9,7 @@ use lemmy_db_schema::{
   source::{
     instance::{InstanceActions, InstanceBanForm},
     local_user::LocalUser,
-    mod_log::moderator::{AdminBan, AdminBanForm},
+    mod_log::admin::{AdminBan, AdminBanForm},
   },
   traits::{Bannable, Crud},
 };
