+++ crates/db_schema/src/impls/mod_log/admin.rs
@@ -1,15 +1,22 @@
 use crate::{
   newtypes::{
+    AdminAddId,
     AdminAllowInstanceId,
+    AdminBanId,
     AdminBlockInstanceId,
     AdminPurgeCommentId,
     AdminPurgeCommunityId,
     AdminPurgePersonId,
     AdminPurgePostId,
+    AdminRemoveCommunityId,
   },
   source::mod_log::admin::{
+    AdminAdd,
+    AdminAddForm,
     AdminAllowInstance,
     AdminAllowInstanceForm,
+    AdminBan,
+    AdminBanForm,
     AdminBlockInstance,
     AdminBlockInstanceForm,
     AdminPurgeComment,
@@ -20,6 +27,8 @@ use crate::{
     AdminPurgePersonForm,
     AdminPurgePost,
     AdminPurgePostForm,
+    AdminRemoveCommunity,
+    AdminRemoveCommunityForm,
   },
   traits::Crud,
   utils::{get_conn, DbPool},
@@ -27,12 +36,15 @@ use crate::{
 use diesel::{dsl::insert_into, QueryDsl};
 use diesel_async::RunQueryDsl;
 use lemmy_db_schema_file::schema::{
+  admin_add,
   admin_allow_instance,
+  admin_ban,
   admin_block_instance,
   admin_purge_comment,
   admin_purge_community,
   admin_purge_person,
   admin_purge_post,
+  admin_remove_community,
 };
 use lemmy_utils::error::{LemmyErrorExt, LemmyErrorType, LemmyResult};
 
@@ -197,6 +209,90 @@ impl Crud for AdminBlockInstance {
   ) -> LemmyResult<Self> {
     let conn = &mut get_conn(pool).await?;
     diesel::update(admin_block_instance::table.find(from_id))
+      .set(form)
+      .get_result::<Self>(conn)
+      .await
+      .with_lemmy_type(LemmyErrorType::CouldntUpdateModlog)
+  }
+}
+
+impl Crud for AdminRemoveCommunity {
+  type InsertForm = AdminRemoveCommunityForm;
+  type UpdateForm = AdminRemoveCommunityForm;
+  type IdType = AdminRemoveCommunityId;
+
+  async fn create(pool: &mut DbPool<'_>, form: &Self::InsertForm) -> LemmyResult<Self> {
+    let conn = &mut get_conn(pool).await?;
+    insert_into(admin_remove_community::table)
+      .values(form)
+      .get_result::<Self>(conn)
+      .await
+      .with_lemmy_type(LemmyErrorType::CouldntCreateModlog)
+  }
+
+  async fn update(
+    pool: &mut DbPool<'_>,
+    from_id: Self::IdType,
+    form: &Self::UpdateForm,
+  ) -> LemmyResult<Self> {
+    let conn = &mut get_conn(pool).await?;
+    diesel::update(admin_remove_community::table.find(from_id))
+      .set(form)
+      .get_result::<Self>(conn)
+      .await
+      .with_lemmy_type(LemmyErrorType::CouldntUpdateModlog)
+  }
+}
+
+impl Crud for AdminBan {
+  type InsertForm = AdminBanForm;
+  type UpdateForm = AdminBanForm;
+  type IdType = AdminBanId;
+
+  async fn create(pool: &mut DbPool<'_>, form: &Self::InsertForm) -> LemmyResult<Self> {
+    let conn = &mut get_conn(pool).await?;
+    insert_into(admin_ban::table)
+      .values(form)
+      .get_result::<Self>(conn)
+      .await
+      .with_lemmy_type(LemmyErrorType::CouldntCreateModlog)
+  }
+
+  async fn update(
+    pool: &mut DbPool<'_>,
+    from_id: Self::IdType,
+    form: &Self::UpdateForm,
+  ) -> LemmyResult<Self> {
+    let conn = &mut get_conn(pool).await?;
+    diesel::update(admin_ban::table.find(from_id))
+      .set(form)
+      .get_result::<Self>(conn)
+      .await
+      .with_lemmy_type(LemmyErrorType::CouldntUpdateModlog)
+  }
+}
+
+impl Crud for AdminAdd {
+  type InsertForm = AdminAddForm;
+  type UpdateForm = AdminAddForm;
+  type IdType = AdminAddId;
+
+  async fn create(pool: &mut DbPool<'_>, form: &Self::InsertForm) -> LemmyResult<Self> {
+    let conn = &mut get_conn(pool).await?;
+    insert_into(admin_add::table)
+      .values(form)
+      .get_result::<Self>(conn)
+      .await
+      .with_lemmy_type(LemmyErrorType::CouldntCreateModlog)
+  }
+
+  async fn update(
+    pool: &mut DbPool<'_>,
+    from_id: Self::IdType,
+    form: &Self::UpdateForm,
+  ) -> LemmyResult<Self> {
+    let conn = &mut get_conn(pool).await?;
+    diesel::update(admin_add::table.find(from_id))
       .set(form)
       .get_result::<Self>(conn)
       .await
