+++ crates/db_views/modlog_combined/src/lib.rs
@@ -4,17 +4,17 @@ use lemmy_db_schema::source::{
   instance::Instance,
   mod_log::{
     admin::{
+      AdminAdd,
       AdminAllowInstance,
+      AdminBan,
       AdminBlockInstance,
       AdminPurgeComment,
       AdminPurgeCommunity,
       AdminPurgePerson,
       AdminPurgePost,
+      AdminRemoveCommunity,
     },
     moderator::{
-      AdminAdd,
-      AdminBan,
-      AdminRemoveCommunity,
       ModAddToCommunity,
       ModBanFromCommunity,
       ModChangeCommunityVisibility,
